+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Update SPM tools-version to 5.3 to enable Swift 5.3 support (#1148)
 - Add backend for swift-log (#1164)
 - Specify CocoaPods version to ensure `swift_version` attribute works (#1167)
+- Simplify `DDLogFileManager` callbacks for archived log files (#1166)
 
 ## [3.6.2 - Xcode 11.6 on July 31st, 2020](https://github.com/CocoaLumberjack/CocoaLumberjack/releases/tag/3.6.2)